added sierpinskiocaml.ml
the ocaml version of my python sierpinski program. significantly faster (well, it goes from 1024 to 1 in 0.5 sec while in the same time the other did 512 to 4), although it leaves weird blank pixels on the edges. it is probably just the graphics module not being made to operate on such a small pixel. after all it looks like turte has some advantages.